restore all bugs

+++ src/Entity/Coupon.php
@@ -21,7 +21,7 @@ class Coupon
     private ?string $discountType = null;
 
     #[ORM\Column]
-    private ?float $value = null;
+    private ?int $value = null;
 
     public function getId(): ?int
     {
@@ -52,12 +52,12 @@ class Coupon
         return $this;
     }
 
-    public function getValue(): ?float
+    public function getValue(): ?int
     {
         return $this->value;
     }
 
-    public function setValue(float $value): static
+    public function setValue(int $value): static
     {
         $this->value = $value;
