+++ src/Entity/Product.php
@@ -18,7 +18,7 @@ class Product
     private ?string $name = null;
 
     #[ORM\Column]
-    private ?float $price = null;
+    private ?int $price = null;
 
     public function getId(): ?int
     {
@@ -37,12 +37,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?float
+    public function getPrice(): ?int
     {
         return $this->price;
     }
 
-    public function setPrice(float $price): static
+    public function setPrice(int $price): static
     {
         $this->price = $price;
